Drastically reduced mem usage
Created a system for handling note lengths without requiring millisecond
time-keeping.
Moved much of the system over to a tick-style time system based on
sixteenth-notes. This simplifies a lot of systems down the line.
Especially in the Pattern class, which no longer needs to keep track of
milliseconds.
The Step class has been greatly simplified, removing all length-timing
stuff, and giving that responsibility to the midi class, a system that
sits outside any pattern stuff (so it always knows which notes are
playing, and can stop them regardless of what happens to the pattern
class that triggered the note.
Changed accumulator types to floats, as Unsigned Longs were causing
clock drift. Floats seem to work perfeclty so far.